New Images, Add min/max levelreq spinboxes, GemList is copyable, Add "all attributes" filter
Needed XCeed WPF Toolkit for NumericUpDown.
Now you can locate items in between certain Level. This only works with
the typed-in search, not the category checkboxes. (little trick, push
Spacebar if you are using category, to force it to search for a space -
most items do have a space, so this will trick/force the Level req  -
will fix later or revamp entirely)
(Min/Max level is kind of clunky, I am bad with MVC style)
The list of gems was using a weird un-copyable richtextbox, I did not
figure out how to re-use the "BindableRichTextBox" already there so I
made a 2nd one that converts strings. Now I can not figure out why it
prints a blank line + \n before the first gem, when the string we get
from GemDistributionstring is completely normal.
Added a way to search for "+All Attributes"