LevelReq Filter now uses a checkbox to enable/disable (instead of the usersearch textbox)
Un-did the changes to the usersearch box. Now "Level Requirement" shows
up as a checkbox in the bottom of the filter list if you want to use the
min/max level req spinboxes. Re-Filters dynamically when the numbers are
modified.
Changed the language of some help tooltips.
Everything else not labeled LEVEL REQUIREMENT CODE is meaningless code
syntax optimization.